+++ app/react/V2/Components/Analitycs/index.ts
